added email length validation

+++ application/source/source/domain_layer/service/_in_context_for/domain_layer/entity/entity/application_user/_new_for_context/validator/mod.rs
@@ -6,7 +6,7 @@ use regex::Regex;
 pub struct Validator;
 
 impl Validator {
-    const EMAIL_MAXIMUM_LENGTH: u8 = 320;
+    const EMAIL_MAXIMUM_LENGTH: u16 = 320;
     const PASSWORD_MAXIMUM_LENGTH: u8 = 7;
 
     pub fn is_valid_email<'outer_a>(email: &'outer_a Email) -> Result<bool, BaseError> {
